Adds marks sensors using coordinator data
Implements new sensors to display new and last marks for each child, utilizing the data from the coordinator.

The old sensors are retained for backward compatibility and will be removed in a future update once the migration is complete.

Also, it changes the order of Library and API version in coordinator to Library: API to be more readable, and removes school from display name in coordinator