Enhances marks data and adds sensors (#71)
* Enhances marks data and adds sensors

Adds a new API endpoint for fetching a compact snapshot of marks, including subjects, grouped marks, and a flattened list of marks.

Implements new sensors to display all marks and aggregate them per subject.

* Adds school year calculation utility

Adds a utility function to calculate the start and end dates of the school year, which will be required for filtering marks for current school year.

* Updates async-bakalari-api to version 0.6.0

Bumps the async-bakalari-api dependency to the latest version (0.6.0).

This ensures compatibility with the newest features and bug fixes in the underlying API library. Also updates the version in the component's const file, manifest file and Makefile.


Improves logging format for better readability and debugging.

* Silence Makefile commands

Adds `@` to the beginning of each line in the Makefile to suppress output during execution.

This change improves readability of the build process output by only showing essential messages.

* Adds automation for Bakaláři new marks notification

Implements a new automation blueprint that sends a notification when new marks are available in Bakaláři.

The automation triggers based on the state of a sensor entity that tracks new marks. It then sends a notification containing details of the latest marks, including the subject, description, date, and mark.